2. BackEnd/1. Micros_lamdas/bp-librz-mto-seguros-voluntarios/src/main/java/com/pichincha/bplibrzmtoparamssegurosvoluntarios/utils/CurrencyDeserializer.java
package com.pichincha.bplibrzmtoparamssegurosvoluntarios.utils;
import com.fasterxml.jackson.core.JsonParser;
import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.DeserializationContext;
import com.fasterxml.jackson.databind.JsonDeserializer;

import java.io.IOException;
import java.math.BigDecimal;

public class CurrencyDeserializer extends JsonDeserializer<BigDecimal> {

    @Override
    public BigDecimal deserialize(JsonParser p, DeserializationContext ctxt)
            throws IOException, JsonProcessingException {
        String value = p.getText();
        value = value.replaceAll("[$,]", "").trim();
        return new BigDecimal(value);
    }
}